Intital Development for EKM Project. Intitlaised Models with a Database First Approach (Nomrally Code first) Established the relationships, which connected them on the pages. Which is awesome!
To do:
Create a SQL Script which will populate all the databasw with TestData.
Create a view model so that a user can see all the books that are associated with a author.
Create a independent table for each user to store a list of books for them.
Implement role managment so that a super admin can add more books, and only them.
Start creating a dynamic front, and try and use ReactJS for a better experince.
Use bootstarp effectivly.
Ensure all secuirty measures are taken: AntiForgery, XXS, Error Prevention

Inform EKM then of the progress.
Try and get done by Friday evening.